+++ azure-communication-ui/azure-communication-ui/src/main/java/com/azure/android/communication/ui/presentation/fragment/factories/SetupViewModelFactory.kt
@@ -3,7 +3,6 @@
 
 package com.azure.android.communication.ui.presentation.fragment.factories
 
-import com.azure.android.communication.ui.configuration.LocalizationProvider
 import com.azure.android.communication.ui.presentation.fragment.common.audiodevicelist.AudioDeviceListViewModel
 import com.azure.android.communication.ui.presentation.fragment.setup.components.ErrorInfoViewModel
 import com.azure.android.communication.ui.presentation.fragment.setup.components.JoinCallButtonHolderViewModel
@@ -18,12 +17,10 @@ import com.azure.android.communication.ui.redux.state.ReduxState
 
 internal class SetupViewModelFactory(
     private val store: Store<ReduxState>,
-    private val localizationProvider: LocalizationProvider,
     private val localParticipantConfig: PersonaManager
 ) {
-
     private val audioDeviceListViewModel by lazy {
-        AudioDeviceListViewModel(store::dispatch, localizationProvider)
+        AudioDeviceListViewModel(store::dispatch)
     }
 
     private val previewAreaViewModel by lazy {
@@ -31,21 +28,15 @@ internal class SetupViewModelFactory(
     }
 
     private val setupControlsViewModel by lazy {
-        SetupControlBarViewModel(
-            store::dispatch,
-            localizationProvider
-        )
+        SetupControlBarViewModel(store::dispatch)
     }
 
     private val warningsViewModel by lazy {
-        PermissionWarningViewModel(
-            store::dispatch,
-            localizationProvider
-        )
+        PermissionWarningViewModel(store::dispatch)
     }
 
     private val snackBarViewModel by lazy {
-        ErrorInfoViewModel(localizationProvider)
+        ErrorInfoViewModel()
     }
 
     private val setupGradientViewModel by lazy {
@@ -57,10 +48,7 @@ internal class SetupViewModelFactory(
     }
 
     private val joinCallButtonHolderViewModel by lazy {
-        JoinCallButtonHolderViewModel(
-            store::dispatch,
-            localizationProvider
-        )
+        JoinCallButtonHolderViewModel(store::dispatch)
     }
 
     fun providePreviewAreaViewModel() = previewAreaViewModel