+++ azure-communication-ui/azure-communication-ui/src/test/java/com/azure/android/communication/ui/presentation/fragment/calling/lobby/LobbyOverlayViewModelTest.kt
@@ -3,8 +3,6 @@
 
 package com.azure.android.communication.ui.presentation.fragment.calling.lobby
 
-import com.azure.android.communication.ui.configuration.AppLocalizationProvider
-import com.azure.android.communication.ui.configuration.LocalizationProvider
 import com.azure.android.communication.ui.helper.MainCoroutineRule
 import com.azure.android.communication.ui.redux.state.CallingStatus
 import kotlinx.coroutines.flow.toList
@@ -21,14 +19,13 @@ internal class LobbyOverlayViewModelTest {
 
     @get:Rule
     var mainCoroutineRule = MainCoroutineRule()
-    private val appLocalizationProvider: LocalizationProvider = AppLocalizationProvider()
 
     @Test
     fun lobbyOverlayViewModel_when_callingStateChange_then_notifyLobbyState() =
         mainCoroutineRule.testDispatcher.runBlockingTest {
 
             // arrange
-            val viewModel = LobbyOverlayViewModel(appLocalizationProvider)
+            val viewModel = LobbyOverlayViewModel()
             viewModel.init(
                 CallingStatus.CONNECTED,
             )